+++ src/views/Chat.vue
@@ -26,4 +26,3 @@ export default class Chat extends Vue {
 <style scoped>
 
 </style>
-
